+++ src/client/core/index.ts
@@ -2,3 +2,4 @@ export * from "./unique-window";
 export * from "./global";
 export * from "./subprocess";
 export * from "./batchlabs-application";
+export * from "./smart-card-certificate";
